Complete project overhaul – rewritten from the base in many areas – made look great on both iOS 6 & 7 – introducing new Weather pane with dynamic background – Re-design & functionality fixes in Events List & Event detail

+++ Hvezdarna/Classes/ProgramList.h
@@ -9,12 +9,13 @@
 #import <Foundation/Foundation.h>
 #import "Program.h"
 
+
 @interface ProgramList : NSObject
 
-- (int) numberOfDays;
-- (int) numberOfEventsOnDayIndex:(NSInteger)idx;
+- (NSUInteger) numberOfDays;
+- (NSUInteger) numberOfEventsOnDayIndex:(NSInteger)idx;
 - (NSInteger) dayAtIndex:(NSInteger)idx;
-- (Program*) programOnDayIdx:(NSInteger)day atIdx:(NSInteger)idx;
-- (void) processSearchWord:(NSString*)word;
+- (Program *) programOnDayIdx:(NSInteger)day atIdx:(NSInteger)idx;
+- (void) processSearchWord:(NSString *)word;
 
 @end
